+++ NEWS.md
@@ -1,8 +1,26 @@
 # Changelog
 
-## 2.42.3
+## 2.44.0
 
 Released 2016-mm-dd
+
+Adds support for sql wrap in all layers
+
+
+## 2.43.1
+
+Released 2016-05-19
+
+Bug fixes:
+ - Dataview error when bbox present without query rewrite data #458
+
+
+## 2.43.0
+
+Released 2016-05-18
+
+New features:
+ - Overviews now support dataviews and filtering #449
 
 
 ## 2.42.2